Move subscription test API to `clientcubscription_test.go` for easy …
